+++ pcs_test/tier1/test_misc.py
@@ -33,10 +33,9 @@ class ParseArgvDashDash(TestCase, AssertPcsMixin):
         self.assert_pcs_fail(
             self.cmd + ["-123"],
             outdent(
-                """\
+                f"""\
                 Deprecation Warning: Using '-123' without '--' is deprecated, those parameters will be considered position independent options in future pcs versions
-                Deprecation Warning: Specifying score as a standalone value is deprecated and might be removed in a future release, use score=value instead
-                Error: Resource 'R1' does not exist
+                Error: invalid role value 'R2', allowed values are: {self.allowed_roles}
                 """
             ),
         )
@@ -67,9 +66,8 @@ class ParseArgvDashDash(TestCase, AssertPcsMixin):
         self.assert_pcs_fail(
             ["--"] + self.cmd + ["-123"],
             outdent(
-                """\
-                Deprecation Warning: Specifying score as a standalone value is deprecated and might be removed in a future release, use score=value instead
-                Error: Resource 'R1' does not exist
+                f"""\
+                Error: invalid role value 'R2', allowed values are: {self.allowed_roles}
                 """
             ),
         )